Set up permissions for commenting

+++ src/WebApp/Pages/Compliance/FCE/Details.cshtml.cs
@@ -39,7 +39,8 @@ public class DetailsModel(IFceService fceService, IAuthorizationService authoriz
             NewComment = new CommentAddDto(Id),
             NewCommentId = NewCommentId,
             NotificationFailureMessage = NotificationFailureMessage,
-            Editable = UserCan[ComplianceWorkOperation.Edit],
+            CanAddComment = UserCan[ComplianceWorkOperation.AddComment],
+            CanDeleteComment = UserCan[ComplianceWorkOperation.DeleteComment],
         };
         return Page();
     }
@@ -51,7 +52,7 @@ public class DetailsModel(IFceService fceService, IAuthorizationService authoriz
         if (item is null || item.IsDeleted) return BadRequest();
 
         await SetPermissionsAsync(item);
-        if (!UserCan[ComplianceWorkOperation.Edit]) return BadRequest();
+        if (!UserCan[ComplianceWorkOperation.AddComment]) return BadRequest();
 
         if (!ModelState.IsValid)
         {
@@ -60,7 +61,8 @@ public class DetailsModel(IFceService fceService, IAuthorizationService authoriz
             {
                 Comments = item.Comments,
                 NewComment = newComment,
-                Editable = UserCan[ComplianceWorkOperation.Edit],
+                CanAddComment = UserCan[ComplianceWorkOperation.AddComment],
+                CanDeleteComment = UserCan[ComplianceWorkOperation.DeleteComment],
             };
             return Page();
         }
@@ -74,6 +76,12 @@ public class DetailsModel(IFceService fceService, IAuthorizationService authoriz
 
     public async Task<IActionResult> OnPostDeleteCommentAsync(Guid commentId, CancellationToken token)
     {
+        var item = await fceService.FindAsync(Id, token);
+        if (item is null || item.IsDeleted) return BadRequest();
+
+        await SetPermissionsAsync(item);
+        if (!UserCan[ComplianceWorkOperation.DeleteComment]) return BadRequest();
+
         await fceService.DeleteCommentAsync(commentId, token);
         return RedirectToPage("Details", pageHandler: null, routeValues: new { Id }, fragment: "comments");
     }
